Flutter app: home_page con 2 tabs, métricas completas + fix CNN datos reales
- home_page.dart: nuevo StatefulWidget con BottomNavigationBar
  Tab 1 (Actividad): actividad actual, confianza, VitaPoints acumulados,
    minutos activos, estado fall monitor en tiempo real, selector de modo
  Tab 2 (Métricas): info de ambos modelos, desglose VitaPoints/actividad,
    cascada 3 etapas explicada, modos conservador/equilibrado/estricto,
    sesión desglosada por actividad, debug en tiempo real
- har_classifier.dart: ActivityInfo extension (displayName, vitaPointsPerMin, isActive)
- fall_detector.dart: eliminar import unused, fix cast innecesario
- main.dart: apunta a HomePage en lugar de ActivityScreen
- notebooks/04_fall_detection.ipynb: fix path SisFall_dataset + loader semicolons
  CNN entrenado con datos reales: Recall 99.7% (balanced), 100% (conservador)
  53592 ventanas reales, 38 sujetos, ratio 12.5:1 ADL/falls

+++ app/lib/inference/fall_detector.dart
@@ -1,5 +1,4 @@
 import 'package:tflite_flutter/tflite_flutter.dart';
-import '../sensors/sensor_service.dart';
 
 class FallResult {
   final int triggeredStage;
@@ -104,7 +103,7 @@ class FallDetector {
     final input = [window.map((s) => s.map((v) => v.toDouble()).toList()).toList()];
     final output = [[0.0]];
     _interpreter!.run(input, output);
-    return (output[0][0] as double).clamp(0.0, 1.0);
+    return output[0][0].clamp(0.0, 1.0);
   }
 
   void dispose() {
